The e-graph walk comes home: luminal::extraction
Phase 1 of the #420/#422 rejoin moved extractor.rs out of core and
duplicated it, verbatim, into every runtime that searches — with the
door left open ("Maybe if there are some core utilities, they can belong
in core"). Seven phases later the three copies had drifted by exactly
one API-shape hunk and zero logic lines: 4,130 of 4,130 shared lines,
and not one `crate::` reference outside the reference copy's tests. The
divergence the ledger said to wait for did not happen.

So the walk is core's again, under the name that says what it is.
`src/extraction.rs` is the CUDA-lite copy — Phase 2's borrowed-matcher
form, the more general one, since a `Vec` owner passes `&vec` — plus the
two test modules the reference copy carried alone (`render_memo_tests`,
`chain_stride_tests`), which now cover the one copy instead of one of
three. `chain_stride_tests` reaches for a registry through core's
`luminal_reference` dev-dependency and spells the crossing `luminal::`,
per the dep-world rule.

Nothing about selection moves. Each runtime still owns its op registry,
its allow list, its evaluator, its option knobs and outcome shape, its
finalist policy, and the loop that runs them; `extractor` in each crate
is now an alias for the one module. Call sites read the same, apart from
an `&` where a runtime handed over an owned matcher list.

test_runtime's genome bridge goes with it: `adopt_genome`/`adopt_choice`
existed only because two structurally identical `Genome` types in two
crates are nominally distinct. There is one `Genome` now.

Zero logic change — the RNG stream, the sampler and both loops are
untouched, and the seeded pins (cublaslt_election, cublaslt_bias_premise,
cublaslt_contracts_cpu, finalists_lattice, dim_buckets) pass unchanged.
